% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS4.R
\name{fars_read}
\alias{fars_read}
\title{R e v i e w  FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Is a character string which represents the archive's name.}
}
\value{
This functions returns a data frame of the imported data in case of an invalid name it will return
        a message error "file XXXX does not exist".
}
\description{
Function #1 This function reads data from .csv file
}
\details{
It is fed with internal data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System, which is a nationwide
census providing the American public yearly data regarding fatal injuries.
suffered in motor vehicle traffic crashes.
}
\examples{
\dontrun{fars_read(filename = "accident_2013.csv.bz2")}


}
