% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS4.R
\name{fars_read_years}
\alias{fars_read_years}
\title{FUN #3 Read FARS years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Could be a single year or a list of years.}
}
\value{
The results of this function will be a list with only the columns MONTH and year, as you can
        confirm watching the select(MONTH, year). If any year in the list was invalid the function show
        a warning message.
}
\description{
Makes a list of month and years
}
\details{
function used by \code{fars_summarize_years}
}
\examples{
\dontrun{fars_read_years(years = list(2013,2014,2015))}


}
