% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS4.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{#'FUN #4 Summarize FARS data by years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Could be a single year or a list pf years.}
}
\value{
The return will be a data.frame with years in columns and months in rows. Each row represent
        the number of accidents.
}
\description{
Makes a summary count per month in particular years.This function summarizes yearly accidents data, by month
}
\examples{
\dontrun{fars_summarize_years(years = list(2013,2014,2015))}


}
