% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS4.R
\name{make_filename}
\alias{make_filename}
\title{FUN #2 Make data file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Could be one number or a list of numbers.}
}
\value{
The return of this function could be a single string or a list, depends of its inputs.
}
\description{
Make .csv data file name related to the given \code{fars_read}
}
\examples{
\dontrun{make_filename(year = 2015)}
\dontrun{make_filename(year = c(2013,2014))}

}
