% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer value which identifies a state in the FARS dataset}

\item{year}{An integer value which for one of the years in the FARS dataset}
}
\value{
A plot of the states map with the coordinates of each fatality that year in
     that state plotted as a point
}
\description{
\code{fars_map_state} returns a plot of the coordinates of each fatal collision by the
 state and year in which they occured
}
\note{
if the state number isnt included in the dataset STATE column, it returns an error
}
\examples{
library(mapdata)
fars_map_state(12, 2014)
fars_map_state(36, 2014)

}
