% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{fars_read}
\alias{fars_read}
\title{fars_read()}
\usage{
fars_read(filename)
}
\arguments{
\item{make_filename}{A string of characters located in WD}
}
\value{
This function returns a tibble of FARS data. As a side
    side effect, it also prints the tibble.
}
\description{
\code{fars_read} reads FARS data .csv and converts it to tibble i.e. dplyr::tbl_df(). it used with
 \code{\link{fars_summarize_years}}, \code{\link{fars_read_years}}, \code{\link{make_filename}}
  and \code{\link{fars_map_state}}.
}
\note{
if filename is not in WD than it returns an error
}
\examples{
fars_read("accident_2015.csv.bz2")

}
