% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{fars_read}
\alias{fars_read}
\title{fars_read()}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string of characters whose data is to become a tibble}
}
\value{
This function returns a tibble with the data from the filname.
}
\description{
\code{fars_read} reads FARS data .csv and converts it to tibble i.e. dplyr::tbl_df(). it will
 return an error if the  \code{filename} is not in the working directory.
}
\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}


}
