% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the numeric vector of years which are being search for.}
}
\value{
returns  tibble of the number of autocollisions in each month over the span
of years
}
\description{
\code{fars_summarize_years}
aggregates the the data into a tibble of the number of autocollisions in each month
spread across the span of years
}
\examples{
fars_summarize_years(years = c(2013, 2014, 2015))
fars_summarize_years(years = 2013)

}
