% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the numeric vector of years which are being search for.}
}
\value{
returns  tibble of the number of autocollisions in each month over the span
of years
}
\description{
A funtion which takes a 4 digit vector \code{years}, reads FARS data into a list of dataframes via
\code{fars_read_years} then aggregates the the data into a tibble of the number of fatal
autocollisions in each month spread across the span of years
}
\examples{
\dontrun{fars_summarize_years(years = c(2013, 2014, 2015))}

}
