% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{fars_summarize_years}{year: an interger value for the year in
the FARS data sourced from the fars_summarize_years function}

\item{fars_map_state}{year: an integer value for year in the FARS data sourced from the mapping
function.}
}
\value{
A character string identical to the filename of the corresponding year in the
    WD
}
\description{
\code{make_file_name} Take as input the year and returns a character string
   that's a filename in the working directory. Used in conjunction with
   \code{\link{fars_summarize_year}} and \code{\link{fars_map_state}}.
}
\examples{
make_filename(2015)

}
