% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a four digit interger}
}
\value{
A character string identical to the filename of the corresponding year in the
    WD
}
\description{
This function takes as input the \code{year} and returns a character string
   that's a filename in the working directory. .
}
\examples{
\dontrun{make_filename(2015)}

}
